% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project-pipeline.R
\name{action_names_to_txt}
\alias{action_names_to_txt}
\title{Put action names in a txt file ----}
\usage{
action_names_to_txt(action_list, filepath = NULL)
}
\arguments{
\item{action_list}{list of project actions}

\item{filepath}{file path and name where .txt file should be saved. If not provided, then prints to console!}
}
\description{
Put action names in a txt file ----
}
\details{
grab all action names and send to a txt file.  "action_list" should be the "actions" list entry in the "project_list" object (i.e., \code{project_list$actions})
}
