% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project-pipeline.R
\name{c_action}
\alias{c_action}
\title{Create comment object}
\usage{
c_action(...)
}
\arguments{
\item{...}{a collection of actions and lists of actions.}
}
\value{
A list of actions.
}
\description{
Create comment object
}
\details{
Use this to combine action objects before passing to \code{project_list()}.
This ensures that the list of actions has the correct structure. Do not use \code{list(...)} or similar!
}
