% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_utils.R
\name{convert_date_to_idnt}
\alias{convert_date_to_idnt}
\title{Title Convert Date to a numeric value}
\usage{
convert_date_to_idnt(date)
}
\arguments{
\item{date}{Vector of dates to be converted to integers}
}
\value{
vector of date idnts
}
\description{
Title Convert Date to a numeric value
}
\examples{
convert_date_to_idnt(Sys.Date())
}
