% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redaction.R
\name{gt_num}
\alias{gt_num}
\title{Convert output of numeric tabulation (redact_summary_num) to gt object}
\usage{
gt_num(x, var_name = "", num_decimals = 1, pct_decimals = 1)
}
\arguments{
\item{x}{The data.frame produced by \code{redact_summary_num}}

\item{var_name}{The variable name}

\item{num_decimals}{Decimal precision for numbers}

\item{pct_decimals}{Decimal precision for percentages}
}
\value{
A gt object
}
\description{
Convert output of numeric tabulation (redact_summary_num) to gt object
}
\details{
This function takes the output of \code{redact_summary_num} and converts it to a gt object (as from the \code{gt} package) for outputting to html/pdf.
}
