% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osutils-package.R
\docType{package}
\name{osutils-package}
\alias{osutils}
\alias{osutils-package}
\title{osutils: Useful Functions for OpenSAFELY}
\description{
Contains functions that are often needed when using the OpenSAFELY platform \url{https://www.opensafely.org/}, such as redaction and low-memory processing.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/wjchulme/osutils}
  \item Report bugs at \url{https://github.com/wjchulme/osutils/issues/}
}

}
\author{
\strong{Maintainer}: William Hulme \email{william.hulme@phc.ox.ac.uk} (\href{https://orcid.org/0000-0002-9162-4999}{ORCID})

Authors:
\itemize{
  \item Tom Palmer \email{remlapmot@hotmail.com} (\href{https://orcid.org/0000-0003-4655-4511}{ORCID})
}

}
\keyword{internal}
