% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project-pipeline.R
\name{pipeline_comment}
\alias{pipeline_comment}
\title{Create comment object}
\usage{
pipeline_comment(...)
}
\arguments{
\item{...}{character or -character-convertible objects}
}
\value{
A list
}
\description{
Create comment object
}
\details{
key:value list element that will be converted to a comment block in yaml when \code{project_list_to_yaml()} is run.
Each comment will be prefixed by "## " and suffixed by " ##".
These comments are first converted to \code{'': '## your comment here ##'} in yaml, and then tidied up to \verb{## your comment here ##} before saving.
}
