% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project-pipeline.R
\name{pipeline_list}
\alias{pipeline_list}
\title{Create entire pipeline list}
\usage{
pipeline_list(..., .version = "3.0", .population_size = 1000L)
}
\arguments{
\item{...}{all actions and comments that go into the entire project pipeline.
These can be provided as a mixture of single actions (from \code{pipeline_action()} function) or as lists of actions (from \code{c_action()} function.)}

\item{.version}{version of opensafely to use}

\item{.population_size}{size of dummy data expectations}
}
\value{
A list
}
\description{
Create entire pipeline list
}
\details{
This function is used to put all actions together in the entire project list,
as well as specifying the project frontmatter (version and expectations).
}
