% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project-pipeline.R
\name{project_list_to_yaml}
\alias{project_list_to_yaml}
\title{Convert list to yaml and save}
\usage{
project_list_to_yaml(project_list, filepath = NULL)
}
\arguments{
\item{project_list}{list object containing all actions (created using action function) and comment-actions (created using comment_action function) and front-matter.}

\item{filepath}{file path and name where yaml file should be saved. If not provided, then prints to console!}
}
\description{
Convert list to yaml and save
}
\details{
Convert list to yaml string and then prints or saves the results. This also does some reformatting of comment blocks, whitespace, etc.
}
