% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saving.R
\name{readtype_delim}
\alias{readtype_delim}
\title{Read a delimited file (including CSV and TSV) into a tibble, and type columns using a separate json file}
\usage{
readtype_delim(
  file,
  suffix = "",
  delim,
  quote = "\\"",
  escape_backslash = FALSE,
  escape_double = TRUE,
  locale = default_locale(),
  na = c("", "NA"),
  quoted_na = TRUE,
  comment = "",
  trim_ws = FALSE
)
}
\arguments{
\item{file}{Delimited file location.}

\item{suffix}{#' @param suffix The suffix used in the name of the json file, which is appended to the delimited file name. Defaults to \code{""} (no suffix), so that the file name is the same as the delimited file name (excluding filetype extensions).}
}
\description{
Read a delimited file (including CSV and TSV) into a tibble, and type columns using a separate json file
}
\details{
Based on the readr::read_delim function. Requires delimited files to be saved using osutils::write_delim, which will also create the json file containing the typing info.
See readr::read_csv for additional parameter details
Datetime and time classes are not supported.
}
