% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redaction.R
\name{redact_tblsummary}
\alias{redact_tblsummary}
\title{Redact tbl_summary object}
\usage{
redact_tblsummary(x, threshold, redact_chr = NA_character_)
}
\arguments{
\item{x}{A tbl_summary object created by the \code{gt} package.}

\item{threshold}{The redaction threshold. All values less than or equal to this threshold will be redacted.}

\item{redact_chr}{The character string used to replace redacted values. Default is "NA".}
}
\value{
A redacted tbl_summary object
}
\description{
Redact tbl_summary object
}
\details{
This function redacts all statistics based on counts less than the threshold (including means, medians, etc)
it also removes potentially disclosive items from the object, namely:
\itemize{
\item \code{x$inputs$data} which contains the input data
\item \code{x$inputs$meta_data} which contains the raw summary table for the table
}
}
