% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redaction.R
\name{redacted_summary_cat}
\alias{redacted_summary_cat}
\title{Summarise a categorical variable and redact if necessary}
\usage{
redacted_summary_cat(
  x,
  threshold = 5L,
  precision = 1L,
  .missing_name = "(missing)",
  .redacted_name = "redacted"
)
}
\arguments{
\item{x}{The vector to summarise and redact.}

\item{threshold}{The redaction threshold. All values less than or equal to this threshold will be redacted (and possibly more; see the \code{\link{redactor}} function)}

\item{precision}{The precision of any rounding that is to be applied to frequency values. Defaults to 1 (no rounding).}

\item{.missing_name}{The string used to replace \code{NA} categories.}

\item{.redacted_name}{The string used to replace redacted values.}
}
\value{
A table of redacted frequencies and proportions.
}
\description{
Summarise a categorical variable and redact if necessary
}
\details{
This function takes a categorical vector (or something that can be coerced to a categorical vector), computes value frequencies and proportions, and redacts according to the rules in \code{\link{redactor}}.
}
