% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redaction.R
\name{redacted_summary_catcat}
\alias{redacted_summary_catcat}
\title{Categorical by categorical cross-tabulation, with redaction if necessary}
\usage{
redacted_summary_catcat(
  x1,
  x2,
  threshold = 5L,
  precision = 1L,
  .missing_name = "(missing)",
  .redacted_name = "redacted",
  .total_name = NULL
)
}
\arguments{
\item{x1}{The first categorical variable.}

\item{x2}{The second categoical variable.}

\item{threshold}{The redaction threshold. All values less than or equal to this threshold will be redacted (and possibly more; see the \code{\link{redactor}} function)}

\item{precision}{The precision of any rounding that is to be applied to frequency values. Defaults to 1 (no rounding).}

\item{.missing_name}{The string used to replace \code{NA} categories.}

\item{.redacted_name}{The string used to replace redacted values.}

\item{.total_name}{The string used to the label the marginal totals. If NULL, no marginal totals are reported.}
}
\value{
A table of redacted frequencies and proportions, arranged in long-format.
}
\description{
Categorical by categorical cross-tabulation, with redaction if necessary
}
\details{
This function takes two categorical vectors (or vectors that can be coerced to a categorical vectors), performs a cross-tabulation, and redacts according to the rules in \code{\link{redactor}}.
proportions are based on x1 totals.
}
