% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redaction.R
\name{redacted_summary_num}
\alias{redacted_summary_num}
\title{Summarise a numeric vector and redact if necessary}
\usage{
redacted_summary_num(x, threshold = 5L, .redacted_name = "redacted")
}
\arguments{
\item{x}{The numeric variable.}

\item{threshold}{The redaction threshold. If the length of \code{x} is less than or equal to this threshold, then no summary values will be reported.}

\item{.redacted_name}{The string used to replace redacted values.}
}
\value{
A table of summary statistics for the variable.
}
\description{
Summarise a numeric vector and redact if necessary
}
\details{
This function takes a numeric vector (or something that can be coerced to one), and summarises it. Summary statistics are redacted according to the rules in \code{\link{redactor}}.
}
