% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codelists.R, R/redaction.R
\name{redactor}
\alias{redactor}
\title{Indicates which values to redact from a vector of frequencies}
\usage{
redactor(n, threshold)

redactor(n, threshold)
}
\arguments{
\item{n}{A vector of integer frequencies or counts from a 1-dimension frequency distribution.}

\item{threshold}{The redaction threshold. All values (and possibly more; see details) less than or equal to this threshold will be redacted.}
}
\value{
A logical vector the same length as \code{n}.

A logical vector the same length as \code{n}.
}
\description{
Indicates which values to redact from a vector of frequencies

Indicates which values to redact from a vector of frequencies
}
\details{
Given a vector of frequencies \code{n}, this function returns a logical vector of frequencies to be redacted.
All frequencies less than or equal to the threshold are redacted.
If the sum the redacted frequencies is also less than or equal to the threshold, then the smallest unredacted frequency is also redacted.

Given a vector of frequencies \code{n}, this function returns a logical vector of frequencies to be redacted.
All frequencies less than or equal to the threshold are redacted.
If the sum the redacted frequencies is also less than or equal to the threshold, then the smallest unredacted frequency is also redacted.
}
