% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redaction.R
\name{redactor2}
\alias{redactor2}
\title{Redact values in a vector based on frequency values}
\usage{
redactor2(n, threshold, x = NULL)
}
\arguments{
\item{n}{A vector of integer frequencies or counts from a 1-dimension frequency distribution.}

\item{threshold}{The redaction threshold. All values (and possibly more; see details) less than or equal to this threshold will be redacted.}

\item{x}{Values to redact. If \code{x} is \code{NULL} then \code{x} redacts values of \code{n}.}
}
\value{
A vector the same length as \code{n}.
}
\description{
Redact values in a vector based on frequency values
}
\details{
If \code{x} is \code{NULL}, then this function redacts values in \code{n} and returns the redacted vector.
If \code{x} is not \code{NULL}, values in \code{x} are redacted according to frequencies in \code{n}.
Values are redacted as follows:
all frequencies less than or equal to the threshold are redacted;
if the sum the redacted frequencies is also less than or equal to the threshold, then the smallest unredacted frequency is also redacted.
#' @export
}
