% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codelists.R
\name{reformat_codelists}
\alias{reformat_codelists}
\title{Converts a json file of codelist names and URLs into an HTML table}
\usage{
reformat_codelists(import_json_from = "./codelists/codelists.json", export_to)
}
\arguments{
\item{import_json_from}{A character containing the path of the json file containing the codelists.
defaults to \code{./codelists/codelists.json} which is the OpenSAFELY standard}

\item{export_to}{The path to which the file should be saved}
}
\description{
Converts a json file of codelist names and URLs into an HTML table
}
\details{
This function currently only exports an HTML file but it can be adapted to output text, markdown, etc.
Ideally this would be an in-build OpenSAFELY feature rather than written externally in R.
}
