% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redaction.R
\name{round_km}
\alias{round_km}
\title{Rounded Kaplan-Meier curves}
\usage{
round_km(data, time, event, strata = NULL, threshold = 6)
}
\arguments{
\item{data}{A data frame containing the required survival times}

\item{time}{Event/censoring time variable, supplied as a character. Must be numeric >0}

\item{event}{Event indicator variables supplied as a character. Censored (\code{0}/\code{FALSE}) or not (\code{1}/\code{TRUE}). Must be logical or integer with values zero or one}

\item{strata}{names of stratification / grouping variables, supplied as a character vector of variable names}

\item{threshold}{Redact threshold to apply}
}
\value{
A tibble with rounded numbers of at risk, events, censored, and derived survival estimates, by strata
}
\description{
Rounded Kaplan-Meier curves
}
\details{
This function rounds Kaplan-Meier survival estimates by delaying events times until at least \code{threshold} events have occurred.
}
