% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{sample_nonoutcomes_n}
\alias{sample_nonoutcomes_n}
\title{Sample patients (or other observational units) based on patient IDs, depending on occurrence of an event or not}
\usage{
sample_nonoutcomes_n(had_outcome, id, n)
}
\arguments{
\item{had_outcome}{A logical indicating if the patient has experienced the outcome or not}

\item{id}{An integer patient identifier with the following properties:
\itemize{
\item consistent between cohort extracts
\item unique
\item completely randomly assigned (no correlation with practice ID, age, registration date, etc etc) which should be true as it based on hash of true IDs
\item strictly greater than zero
}}

\item{n}{The number of patients (amongst all those who did not experience the event) to be sampled}
}
\value{
A logical vector indicating whether the patient has been sampled or not
}
\description{
Sample patients (or other observational units) based on patient IDs, depending on occurrence of an event or not
}
\details{
If \code{had_outcome} is \code{TRUE} then result is always \code{TRUE}.
If \code{had_outcome} is \code{FALSE}, then result is \code{TRUE} with probability \code{max(1,n/sum(1-had_outcome))} and \code{FALSE} with probability \code{min(0, 1 - n/sum(1-had_outcome))}.
Patients are selected in ascending order of patient ID until the sampling number is met.
Warns (does not fail) if \code{n} is greater than \code{sum(1-had_outcome)}.
}
