% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{sample_random_n}
\alias{sample_random_n}
\title{Sample n patients (or other observational units) based on patient IDs.}
\usage{
sample_random_n(id, n)
}
\arguments{
\item{id}{An integer patient identifier with the following properties:
\itemize{
\item consistent between cohort extracts
\item unique
\item completely randomly assigned (no correlation with practice ID, age, registration date, etc etc) which should be true as it based on hash of true IDs
\item strictly greater than zero
}}

\item{n}{The number of patients (amongst all those who did not experience the event) to be sampled}
}
\value{
A logical vector indicating whether the patient has been sampled or not
}
\description{
Sample n patients (or other observational units) based on patient IDs.
}
\details{
Result is \code{TRUE} with probability \code{max(1,n/length(id))} and \code{FALSE} with probability \code{min(0, 1 - n/length(id))}.
Patients are selected in ascending order of patient ID until the sampling number is met.
Warns (does not fail) if \code{n} is greater than \code{length(id)}.
}
