% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{sample_weights}
\alias{sample_weights}
\title{Derive sampling probabilities}
\usage{
sample_weights(had_outcome, sampled)
}
\arguments{
\item{had_outcome}{A logical indicating if the patient has experienced the outcome or not}

\item{sampled}{A logical indicating if a patient was sampled or not}
}
\value{
A numeric vector of the sampling probability
}
\description{
Derive sampling probabilities
}
