% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{split_train_dt}
\alias{split_train_dt}
\title{Title}
\usage{
split_train_dt(train_dat, partitions = 3, prop = c(0.6, 0.2, 0.2), seed = 123)
}
\arguments{
\item{train_dat}{Dataframe to be split}

\item{partitions}{Number of partitions that the data is to be split into (defaults to 3)}

\item{prop}{Set of proportions to split the dataset (should sum to 1)}

\item{seed}{}
}
\value{
list of split datasets
}
\description{
Title
}
