% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saving.R
\name{writetype_csv}
\alias{writetype_csv}
\title{Write a data frame to a csv file, and save typing information in a separate json file}
\usage{
writetype_csv(
  x,
  path,
  suffix = "",
  na = "NA",
  quote_escape = "double",
  eol = "\\n"
)
}
\arguments{
\item{x}{A data frame or tibble to write to disk.}

\item{path}{File or connection to write to. (path is now deprecated in readr v1.4 for OpenSAFELY currently has older version, so use path for now).}

\item{suffix}{The suffix used in the name of the json file, to be appended to the delimited file name. Defaults to \code{""} (no suffix), so that the file name is the same as the delimited file name (excluding filetype extensions).}

\item{na}{String used for missing values. Defaults to NA. Missing values will never be quoted; strings with the same value as na will always be quoted.}

\item{quote_escape}{The type of escaping to use for quoted values, one of "\code{double}", "\code{backslash}" or "\code{none}". You can also use \code{FALSE}, which is equivalent to "\code{none}". The default is "\code{double}", which is expected format for Excel.}

\item{eol}{The end of line character to use. Most commonly either "\verb{\\n}" for Unix style newlines, or "\verb{\\r\\n}" for Windows style newlines.}
}
\value{
Returns the input invisibly
}
\description{
Write a data frame to a csv file, and save typing information in a separate json file
}
\details{
Based on the readr::write_delim function.
Additionally, this function saves a json file containing typing info for the data frame,
which can be used to re-type the data when re-imported into R.
Datetime and time classes are not supported.
}
